requirements.txt
openai>=1.35.0
anthropic>=0.30.0
streamlit>=1.35.0
chromadb>=0.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Gmail OAuth2 (for jcw.loong@gmail.com)
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
